CSV Parser Connor Berry
# CSVMS3
Author Connor Berry
# About
This is a CSV Parser that takes a csv file, and separates the contents into 2 different files, 1, an sqlite database file and 2, another CSV file which contains failed or incorrect records.
# Setup
In order to setup this project, download from Github and import it into Eclipse. WWarning: May be incompatible with versions earlier than Version 2019-12 (4.14.0).
Prepare the proper CSV File, this project accepts 10 columns exactly.
Run the Main file.
The file is now setup.
# Prerequisites to recreation
In order to replicate this project, 3 external objects are needed, Gradle, Maven, & SQLite.
# References
https://bitbucket.org/xerial/sqlite-jdbc/downloads/ to download SQLITE
https://www.tutorialspoint.com/sqlite/sqlite_java.htm to understand the installation of SQLITE
https://maven.apache.org/install.html installing Maven
https://gradle.org/install/ installing Gradle